[Finishes #162265094] updated app routes
refactor views and models
