+++ app/api/v1/users/models.py
@@ -1,6 +1,7 @@
 from app.api.v1.common.base_model import Model
 
+user_list = []
 
 class UserModel(Model):
-    def __init__(self, user_list):
+    def __init__(self):
         super().__init__(user_list)
